API Permissions Table Update
I was working through this procedure yesterday with a customer and we attempted to use the "Delegated" permissions type.  This did not work, but when we changed it to "Application" everything began working.  I have removed the "Delegated or" from the User and Device rows on the table.  I have also removed the rows that said "Unsupported" for personal Microsoft accounts and added a blurb under the table.  Inclusion in the table just confused things, IMO.