Update aws-sources.md (#1396)
* Update aws-sources.md

There is also an issue on lines 80 to 82, it is not formatting properly on the page itself.  I'm not sure why it is behaving this way.  It should have the green window background and "INFO" but it does not render that way on the help page itself. It should look like how the text on lines 99 - 101 renders. On the help page itself it displays as plain text, ":::important Sumo Logic highly recommends using an SNS Subscription Endpoint for its ability to maintain low-latency collection. This is essential to support up-to-date Alerts. :::"

* Update aws-sources.md

up-to-date

* Update docs/send-data/hosted-collectors/amazon-aws/aws-sources.md

* Update docs/send-data/hosted-collectors/amazon-aws/aws-sources.md

* Update docs/send-data/hosted-collectors/amazon-aws/aws-sources.md

* Update docs/send-data/hosted-collectors/amazon-aws/aws-sources.md

* Update docs/send-data/hosted-collectors/amazon-aws/aws-sources.md

* Update docs/send-data/hosted-collectors/amazon-aws/aws-sources.md

Co-authored-by: Stacy Kornluebke <stacy@sumologic.com>
Co-authored-by: Megha Jain <116072777+JainM6@users.noreply.github.com>