Update cloud-soar-integration-framework.md
incidentsDetail is passed to the trigger action for an incidentCustomAction. But your documentation in this location doesn't indicate that it was passed.

Later on in your documentation you have an example trigger yaml that shows it being passed and parsed in the python script. https://help.sumologic.com/docs/cloud-soar/cloud-soar-integration-framework/#trigger-incidentcustomaction-and-taskcustomaction